Add default chains storage management for desktop and extension clients
Introduce a new defaultChains storage module to handle default chain retrieval, improving code organization and simplifying the core storage interface. Migrate existing default chain functions from the core storage and remove obsolete implementations from the extension client.
