--- core/ui/vault/state/defaultChains.ts
@@ -1,9 +1,0 @@
-import { Chain } from '@core/chain/Chain'
-
-export const initialDefaultChains = [
-  Chain.Bitcoin,
-  Chain.Ethereum,
-  Chain.THORChain,
-  Chain.Solana,
-  Chain.BSC,
-]
